Move fetch code into it's own file
